airflow-docker/requirements.txt
apache-airflow-providers-myslq==4.0.0
pandas==1.0.2